updated db schema + jpa entities

+++ RemoteServer/src/main/java/com/example/PlaceFinder/entity/Room.java
@@ -20,7 +20,7 @@ public class Room implements Serializable {
     private String idRoom;
 
     private int numSeats;
-    private int capacity;
+    private float capacity;
 /*
     @OneToMany(mappedBy = "room") //references room field in the Slot class
     private List<Slot> slots = new ArrayList<>();
